suppress few trivial warnings

+++ cpp/include/raft/comms/test.hpp
@@ -230,7 +230,7 @@ bool test_pointToPoint_simple_send_recv(const handle_t &h, int numTrials) {
     for (int printrank = 0; printrank < communicator.get_size(); ++printrank) {
       if (communicator.get_rank() == printrank) {
         std::cout << "Rank " << communicator.get_rank() << " received: [";
-        for (int i = 0; i < received_data.size(); i++) {
+        for (size_t i = 0; i < received_data.size(); i++) {
           auto rec = received_data[i];
           std::cout << rec;
           if (rec == -1) ret = false;
